Move response part of the client transaction into a separate file
