+++ neqo-http3/src/lib.rs
@@ -15,6 +15,7 @@ mod control_stream_local;
 mod control_stream_remote;
 pub mod hframe;
 mod hsettings_frame;
+mod response_stream;
 pub mod server;
 mod server_connection_events;
 mod server_events;
